Add SmallPlanCard widget for displaying budget progress in WalletScreen
